refactor: WinningRank 패키지 변경
