+++ src/main/java/lotto/core/enums/WinningRank.java
@@ -1,4 +1,4 @@
-package lotto.core.model;
+package lotto.core.enums;
 
 import java.math.BigDecimal;
 import java.text.DecimalFormat;
