+++ src/main/java/lotto/core/model/LottoResult.java
@@ -4,6 +4,7 @@ import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.List;
 import lotto.commons.util.Collections;
+import lotto.core.enums.WinningRank;
 
 public class LottoResult {
