+++ src/test/java/lotto/core/model/LottoTicketTest.java
@@ -1,6 +1,7 @@
 package lotto.core.model;
 
 import java.util.List;
+import lotto.core.enums.WinningRank;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
